# docs/api/jxh-manager-openapi.yaml
openapi: 3.1.0
info:
  title: Jxh Manager Admin API
  version: 1.0.0
  description: |-
    Jxh Manager 管理面板与 Jxh-Go 同进程管理端之间的 HTTP 契约。
    所有资源 ID、QQ、群号和外部申请标识均使用字符串编码。
    尚未完成的接口通过 x-stage 和 x-status 标记，不表示当前服务已经实现。
  license:
    name: Private
    identifier: LicenseRef-Private
servers:
  - url: /api/admin/v1
    description: 相对当前管理端 Origin 的 API 基础路径
...
security:
  - cookieAuth: []
paths:
...
  /analytics/summary:
    get:
      tags: [Analytics]
      summary: 获取统计汇总
      operationId: getAnalyticsSummary
      x-stage: phase_2
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/AnalyticsFrom'
        - $ref: '#/components/parameters/AnalyticsTo'
        - $ref: '#/components/parameters/AnalyticsGroupIds'
        - $ref: '#/components/parameters/AnalyticsFeatureKey'
        - $ref: '#/components/parameters/AnalyticsResult'
        - $ref: '#/components/parameters/AnalyticsTimezone'
      responses:
        '200':
          description: 指标值、前周期比较和数据新鲜度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /analytics/timeseries:
    get:
      tags: [Analytics]
      summary: 获取统计时间序列
      operationId: getAnalyticsTimeseries
      x-stage: phase_2
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/AnalyticsFrom'
        - $ref: '#/components/parameters/AnalyticsTo'
        - $ref: '#/components/parameters/AnalyticsGroupIds'
        - $ref: '#/components/parameters/AnalyticsFeatureKey'
        - $ref: '#/components/parameters/AnalyticsResult'
        - $ref: '#/components/parameters/AnalyticsTimezone'
        - name: granularity
          in: query
          required: true
          schema:
            type: string
            enum: [hour, day]
        - name: metric
          in: query
          required: true
          style: form
          explode: true
          schema:
            type: array
            minItems: 1
            maxItems: 10
            items:
              $ref: '#/components/schemas/AnalyticsMetricKey'
      responses:
        '200':
          description: 按粒度聚合的一个或多个指标序列
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsTimeseries'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /analytics/rankings:
    get:
      tags: [Analytics]
      summary: 获取统计排行
      operationId: getAnalyticsRankings
      x-stage: phase_2
      x-status: implemented
      parameters:
        - $ref: '#/components/parameters/AnalyticsFrom'
        - $ref: '#/components/parameters/AnalyticsTo'
        - $ref: '#/components/parameters/AnalyticsGroupIds'
        - $ref: '#/components/parameters/AnalyticsFeatureKey'
        - $ref: '#/components/parameters/AnalyticsResult'
        - $ref: '#/components/parameters/AnalyticsTimezone'
        - name: dimension
          in: query
          required: true
          schema:
            type: string
            enum: [group, command, knowledge_entry]
        - name: metric
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/AnalyticsMetricKey'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: 稳定 key、展示名称、值和排名
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsRankings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jxh_admin_session
      description: HttpOnly 服务端会话 Cookie；响应体不会返回其值。
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 登录或 /auth/me 返回、仅由前端保存在内存中的 CSRF Token。
  headers:
    RequestId:
      description: 用于关联 API、审计和上游调用日志的请求 ID
      schema:
        $ref: '#/components/schemas/Identifier'
...
  parameters:
    Page:
      name: page
      in: query
      description: 从 1 开始的页码；大于 1 时不得与 cursor 同时传入
      schema:
        type: integer
        minimum: 1
        maximum: 100000
        default: 1
...
    AnalyticsFrom:
      name: from
      in: query
      description: 聚合起点；省略时由服务端按接口默认时间范围计算
      schema:
        type: string
        format: date-time
    AnalyticsTo:
      name: to
      in: query
      description: 聚合终点，不得早于 from
      schema:
        type: string
        format: date-time
    AnalyticsGroupIds:
      name: group_id
      in: query
      description: 可重复传入多个群号
      style: form
      explode: true
      schema:
        type: array
        maxItems: 100
        uniqueItems: true
        items:
          $ref: '#/components/schemas/Identifier'
    AnalyticsFeatureKey:
      name: feature_key
      in: query
      style: form
      explode: true
      schema:
        type: array
        uniqueItems: true
        items:
          $ref: '#/components/schemas/FeatureKey'
    AnalyticsResult:
      name: result
      in: query
      style: form
      explode: true
      schema:
        type: array
        uniqueItems: true
        items:
          type: string
          enum: [success, failed, denied, unknown, fallback, skipped]
    AnalyticsTimezone:
      name: timezone
      in: query
      description: IANA 时区名
      schema:
        type: string
        maxLength: 64
        default: Asia/Shanghai
  responses:
    BadRequest:
      description: 请求语法、字段或筛选条件无效
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: validation_failed
              message: 请求字段校验失败
              request_id: req_01J4ZP8Y
              fields:
                name: [格式不正确]
              retryable: false
    Unauthorized:
      description: 未登录、会话过期或会话已撤销
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthenticated
              message: 请重新登录
              request_id: req_01J4ZP8Y
              fields: {}
              retryable: false
    Forbidden:
      description: CSRF 校验失败或当前角色无权执行操作
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: permission_denied
              message: 当前账号无权执行此操作
              request_id: req_01J4ZP8Y
              fields: {}
              retryable: false
...
  schemas:
    Identifier:
      type: string
      minLength: 1
      maxLength: 256
      description: 不透明字符串标识；客户端不得推断前缀、长度或编码含义。
...
    Timestamp:
      type: string
      format: date-time
      description: UTC RFC 3339 时间点
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id, fields, retryable]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
          description: 稳定机器错误码；客户端必须为未知值提供通用降级显示。
        message:
          type: string
          maxLength: 500
        request_id:
          $ref: '#/components/schemas/Identifier'
        fields:
          type: object
          description: 字段路径到一条或多条安全校验消息的映射
          additionalProperties:
            type: array
            items:
              type: string
              maxLength: 300
        retryable:
          type: boolean
...
    FeatureKey:
      type: string
      enum: [keyword_reply, ai_qa, quote, link_cleaner, welcome, custom_commands]
...
    AnalyticsMetricKey:
      type: string
      enum:
        - keyword_reply_count
        - knowledge_trigger_count
        - ai_request_count
        - ai_success_rate
        - ai_duration_ms
        - join_request_count
        - manual_approval_count
        - automatic_approval_count
        - scheduled_job_run_count
        - group_message_count
        - command_run_count
        - active_user_count
        - link_clean_count
        - quote_success_count
        - quote_fallback_count
        - quote_failure_count
    AnalyticsUnit:
      type: string
      enum: [count, percent, milliseconds]
    AnalyticsWindow:
      type: object
      additionalProperties: false
      required: [from, to, timezone]
      properties:
        from:
          $ref: '#/components/schemas/Timestamp'
        to:
          $ref: '#/components/schemas/Timestamp'
        timezone:
          type: string
          maxLength: 64
    AnalyticsMetric:
      type: object
      additionalProperties: false
      required: [key, label, unit, available, value, previous_value, change_percent]
      properties:
        key:
          $ref: '#/components/schemas/AnalyticsMetricKey'
        label:
          type: string
          maxLength: 100
        unit:
          $ref: '#/components/schemas/AnalyticsUnit'
        available:
          type: boolean
        value:
          type: [number, 'null']
        previous_value:
          type: [number, 'null']
        change_percent:
          type: [number, 'null']
    AnalyticsSummary:
      type: object
      additionalProperties: false
      required: [window, metrics, data_fresh_at]
      properties:
        window:
          $ref: '#/components/schemas/AnalyticsWindow'
        metrics:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsMetric'
        data_fresh_at:
          $ref: '#/components/schemas/Timestamp'
    AnalyticsPoint:
      type: object
      additionalProperties: false
      required: [bucket_start, value]
      properties:
        bucket_start:
          $ref: '#/components/schemas/Timestamp'
        value:
          type: [number, 'null']
    AnalyticsSeries:
      type: object
      additionalProperties: false
      required: [metric, label, unit, points]
      properties:
        metric:
          $ref: '#/components/schemas/AnalyticsMetricKey'
        label:
          type: string
          maxLength: 100
        unit:
          $ref: '#/components/schemas/AnalyticsUnit'
        points:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsPoint'
    AnalyticsTimeseries:
      type: object
      additionalProperties: false
      required: [window, granularity, series, data_fresh_at]
      properties:
        window:
          $ref: '#/components/schemas/AnalyticsWindow'
        granularity:
          type: string
          enum: [hour, day]
        series:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsSeries'
        data_fresh_at:
          $ref: '#/components/schemas/Timestamp'
    AnalyticsRankingItem:
      type: object
      additionalProperties: false
      required: [key, display_name, value, rank]
      properties:
        key:
          $ref: '#/components/schemas/Identifier'
        display_name:
          type: string
          maxLength: 200
        value:
          type: number
        rank:
          type: integer
          minimum: 1
    AnalyticsRankings:
      type: object
      additionalProperties: false
      required: [window, dimension, metric, unit, items, total_count, data_fresh_at]
      properties:
        window:
          $ref: '#/components/schemas/AnalyticsWindow'
        dimension:
          type: string
          enum: [group, command, knowledge_entry]
        metric:
          $ref: '#/components/schemas/AnalyticsMetricKey'
        unit:
          $ref: '#/components/schemas/AnalyticsUnit'
        items:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsRankingItem'
        total_count:
          type: integer
          minimum: 0
        data_fresh_at:
          $ref: '#/components/schemas/Timestamp'
...